refactor(common): 提取富文本 HTML 净化器到 common 应用
把 news/serializers.py 的 _sanitize_content 及其白名单（标签/属性/协议/视频嵌入域名）
抽到新 common 应用为 sanitize_html，news 改为引用；SanitizeVideoTest 随之迁入
common/tests 并扩充（新增脚本/事件处理器/javascript: 协议清洗用例）。行为零变更。

为 about 页富文本字段复用同款净化铺路（#22）。
