+++ common/__init__.py
@@ -1,0 +1,1 @@
+"""跨应用共享的工具与基元（无模型、无迁移的轻量 app）。"""
