+++ common/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class CommonConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "common"
+    verbose_name = "通用"
