+++ config/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
     'corsheaders',
     'rest_framework',
     'django_filters',
+    'common',
     'accounts',
     'tasks',
     'messaging',
